feat(playground): Allow specifying a JSON string as the linter config (#11710)
When creating bug reports it would be helpful to have a link to a
reproduction that just worked in the browser. By adding the ability to
config Oxlint here, simple reproductions can be done within the
Playground and URLs shared from it.

This would be a lot better if the `config` was a JSON object instead of
a string, but I don't know how to do that only on the Rust side.

An option would be to keep the Rust side a string, but on the browser
side allow it to be an object. Whenever the object changes, we serialize
that to a string before sending it to the Rust side. If there's interest
in this, then I think I can make this change in the playground repo.

Example config and screenshots below.

```json
"linter": {},
```

![image](https://github.com/user-attachments/assets/177cdb75-4a45-4c21-bf8b-2e7f46af56df)

```json
"linter": {
  "config": "{\"plugins\":[\"typescript\"],\"categories\":{\"correctness\":\"error\"},\"rules\":{\"no-debugger\":\"warn\"}}"
},
```

![image](https://github.com/user-attachments/assets/7a0843c3-d5e3-4002-bb93-964ddcd0629a)