refactor(linter): remove `RulesCache` (#11981)
Keeping `RulesCache` makes `ConfigStoreBuilder` wayy too complicated.
For now, we are going to remove it (ther may be a tiny perf hit, but in the majority of repos, this code is ony run once so should be ok.
Once custom JS plugins are complete, we can consider adding back in the cache