feat(minifier): implement `Number` known methods (#12078)
My vacation is near, so I may have some time resuming my contributions.
Here it is just implementing some functionalities that may never be used
in actual usage :-)

I'm not sure if the minsize result is accurate, since there may be a
problem with my `minsize` locally. I'll check my local `minsize` later,
and I'll use the CI as a test this time.

---------

Co-authored-by: Boshen <boshenc@gmail.com>