feat(napi/oxlint): read source text into start of allocator (#12122)
In `napi/oxlint2`, read source text into start of allocator, instead of end. This is what raw transfer needs.

The implementation here is a complete dog's dinner! In order to inject this code into `oxlint` library code without affecting the behavior of existing `oxlint` app, it uses `oxlint2` feature to changes the allocator's behavior.

There is also way too much unsafe code for comfort. Much of the code in this PR can be removed again when we replace `bumpalo` with our own allocator, because that allocator will allocate all strings at start of the arena by default anyway.

So I propose that we accept this mess as a temporary solution just to get us to a working prototype, and then clean it up in next couple of weeks.