ci(test): run tests on big endian with `--all-features` (#12167)
Big endian tests were failing on CI. This is due to a strange combination of cargo features getting enabled in the tests, causing problems in linter crates, since #12122.

I don't entirely understand how this was happening, but running the tests with `cross test --all-features` fixes it. In any case, we do want to run tests for all features on big endian.