perf(transformer/styled_components): replace hashmap with array (#12170)
Follow-on after #12066.

`helpers` was an `FxHashMap<String, SymbolId>`, but we know all the possible keys, and there's only 6 of them. So use a small array `[Option<SymbolId>; 6]` instead, and convert from strings to indexes into that array via a `StyledComponentsHelper` enum.

Array (24 bytes) is smaller than an `FxHashMap`, and involves no allocations, as well as avoiding hashing strings.

Note: Using an enum rather than an untyped `usize` index, as with an enum, compiler can statically prove indexing into the array can never be out of bounds, so it skips bounds checks.