refactor(transformer/styled_components): make unreachable branch `unreachable!` (#12172)
This branch is unreachable, so use `unreachable!` for it, rather than `return`.

Despite `unreachable!` producing more code, this may still be more performant, because it tells the compiler which branch is going to be taken. Compiler will generate assembly to make the branch which is taken the fast path. Either way, it's a micro micro optimization.

Mainly I think it's preferable because it's clearer when reading the code.