docs(transformer/styled-components): add comments about `CSSMinifier` (#12197)
Added some comments to describe `CSSMinifier`, and clarify why `quasis.set_len(expressions.len() + 1);` is safe.

Also, there is a `assert` in codegen to ensure that `quasis` is always one greater than `expressions`

https://github.com/oxc-project/oxc/blob/c003ebad60970d27aae48b488ff361ef2b612379/crates/oxc_codegen/src/gen.rs#L2091-L2092

In addition, I found that we can use `truncate()` instead of `set_len`. Both are the same, but `truncate` is a safe function.