fix(allocator)!: remove `vec!` macro (#12206)
Remove `vec!` macro from `oxc_allocator`. This is in the `vec2::raw_vec` module, but is accidentally exposed due to the `#[macro_export]` attribute.

This macro creates a `bumpalo::collections::Vec`, instead of our own `Vec` type, which is confusing.

We could add a `vec!` macro which creates our own `Vec` type, but that's not completely trivial (how to handle cloning `item` in `vec![item; 3]`? `Clone` or `CloneIn`?). So for now, just remove this erroneous macro entirely.