perf(semantic,linter): simplify implementation and uses of ancestors iterators (#12164)
This simplifies the implementation and uses of `ancestor_ids`,
`ancestor_kinds` and `ancestors` iterators by changing its first
returned item from the provided node to the first parent of this node.
This makes the implementation of those iterators simpler and makes code
in most lint rules shorter by mainly removing the often called
`.skip(1)`.

All tests are green, but as always I probably broke something untested
because of the scale and impact of this change. I'm not quite happy with
me having to resort to `iter::once` twice in `no_unused_vars` lint rule,
but I couldn't come up with a better solution. But having much less
code, simpler implementation of the iterators and hopefully a
performance gain in the linter benchmarks is definitely worth the
change.

Continuation of #12123 and #12087
Closes #12151