refactor(transformer/styled-components): store block name in arena (#12216)
In some circumstances, `block_name` can be inserted into AST in many places. So put it in arena, and reuse the same string over and over, rather than creating a new arena string each time.

Sorry @Dunqing! It was me that said store it as a `String`. I was wrong!