fix(linter): Report implicit config parse errors (#12260)
Fixes #12258 

Not super sure about the config path normalization, but it Works On My
Machine™️! Only other example of `<cwd>` in linter snapshots is the
`tsconfig` error, which explicitly formats the error at that location.

Might be nicer to regex the snapshot for any paths to handle Windows,
the current cwd detection in there seems a bit hacky anyway.

---------

Co-authored-by: Cameron Clark <cameron.clark@hey.com>