perf(napi/parser): raw transfer: move check for supported platform (#12274)
Small optimization to raw transfer. Move check that running on a supported platform out of `prepareRaw` function (which runs on every call to `parseSync`) to top level (so we only do the check once). This module is lazy-loaded if user calls `parseSync` with `experimentalRawTransfer`, so the check won't run if user is not using raw transfer.