perf(transformer/styled-components): remove quasis and expressions in batch (#12256)
Remove `quasis` and `expressions` in batch instead of removing them one by one, so that it would be more efficient. Considering that the expression's removal is unlikely to be common, I add a `has_deletion_marker` flag to avoid unnecessary `retain` calls, so that we can keep the performance as before in this situation.