feat(allocator): `FixedSizeAllocator` store flag recording if owned by both Rust and JS (#12381)
Add a field `is_double_owned` to `FixedSizeAllocatorMetadata`. This field tracks ownership of the allocator/buffer. It's `true` if both Rust and JS currently hold a reference to the memory, or `false` if only one does.

`Drop` impl for `FixedSizeAllocator` checks the value of this field and only frees the memory if JS doesn't also hold a reference.

At present, `is_double_owned` is always `false`, because we're not transferring buffers over to JS yet. But it'll come into play in later PR. At that point the finalizer which runs when JS garbage collector collects the buffer will also update `is_double_owned`, and free the memory if Rust already dropped the `FixedSizeAllocator`, so JS is the only remaining owner.