refactor(linter, napi/parser): add `source_len` field to `RawTransferMetadata` (#12383)
Add a field `source_len` to `RawTransferData`, which stores the byte length of source text.

In `napi/parser`, this wasn't required, as the source text is passed in from JS side, so JS already knows how long it is. But in linter, source text passes in the opposite direction, so Rust needs to tell JS it's length. It will do so by recording it in this field, and JS reads it from the buffer.