feat(napi/oxlint): pass AST in buffer to JS (#12350)
First working prototype of JS linter plugins.

Rust now passes the AST to JS via buffer, and lint rule runner uses the JS rule's provided visitor to traverse the AST using lazy deserialization.

Only a single rule is supported at present.

Before anyone gets too excited, this is very much unfinished! It's incomplete, the AST is not compliant with ESTree, and it's not well tested. Really this is a proof of concept more than a usable feature at present. But we'll get there...