perf(napi/parser, napi/oxlint): lazy visit: faster check for exit visitor (#12496)
In the absence of benchmarks, I don't know for sure that this is a perf improvement. But I assume so.

`if (exit)` requires coercion of `exit` to a boolean, whereas `if (exit !== null)` is a simpler check - comparison to a single known constant.

Once we have reliable benchmarking set up for raw transfer, we can revisit and tweak the code in many ways, but I think in the meantime this change is *probably* an improvement.