perf(formatter): optimize `SimpleArgument` (#12585)
Optimize the previous implementation, which was ported from Biome, and Biome was ported from `Prettier`. The implementation does not fit with `Rust`, which produces many unnecessary repetitive checks again and again

The new implementation moved most of the checks into a `match` pattern, allowing us to reduce the extra check overhead in each check function. Also, the new implementation purpose was prepared for `checking` ChainExpression, see #12585