refactor(ast)!: introduce `WithClauseKeyword` (#12741)
`WithClause` keyword can only be one of 2 words - `with` or `assert`.

We don't need `Span` for this symbol, so using `IdentiferName` (24 bytes) is overkill. Just use an enum `WithClauseKeyword` (1 byte).

This also has the benefit of constraining the keyword to valid values - prevents creating an invalid AST which would be printed as e.g. `import foo from "source" TROUSERS {};`.