chore(ci): remove paths ignore list in CI task (#12749)
Remove the `paths-ignore` list from `CI` job.

It's ideal to skip running CI when a PR only touches `.md` files. But problem is that these jobs are marked as "required", so CI task *has* to run, or the PR can't be merged.

Maybe we can remove the requirement that these jobs complete before a PR can be merged, but for now this is the best solution.