perf(parser): register `import` / `export` statements in module record directly (#12807)
Currently when parsing a list of statements in `parse_directives_and_statements`, we:

1. Check on every turn of the loop whether the statement is a `ModuleDeclaration`, and whether at top level.
2. If so, we call `ModuleRecordBuilder::visit_module_declaration`.
3. `visit_module_declaration` branches again on the type of the `ModuleDeclaration`.

Instead, remove `ModuleRecordBuilder::visit_module_declaration` method, and instead call individual methods of `ModuleRecordBuilder` for each type of declaration directly, when parsing that particular declaration.

This removes the need to check *every* statement in the loop in `parse_directives_and_statements` just in case it's a module declaration.

Gives a small perf boost (+0.2%) on parser benchmarks. Might be a little bit more in real world, as this removes an unpredictable branch, which CodSpeed doesn't measure.