fix(linter): resolve configured rules for every file linted by `tsgolint` (#12886)
The previous implementation of config resolution was simply incorrect. It did not resolve rule severities and overrides and nested configs properly. Now, although inefficient, we resolve the config file for every path that we will lint with `tsgolint` to correctly account for any possible nested configs, overrides, and so on.