refactor(linter): reduce repeated code (#13070)
Refactor. Many branches of `process_path` return a default `ProcessedModule`. All branches return a `ModuleProcessOutput` with `path` field containing a clone of `path` param.

Move the duplicated logic into a separate function, to allow shortening and simplifying the logic.

There may be a small perf benefit of not repeating the same code in many places, but mainly the motivation is to shorten the code, which I find makes it a bit easier to understand.