feat(linter): implement fixer for jsx_curly_brace_presence (#13005)
This is my first rust code, so an expert review might be appreciated.
Especially the `report_missing_curly_for_text_node` method could use a
good look. I am not sure I made the right choices when it comes to
performance as I miss a lot of the nuance rust offers.

---------

Co-authored-by: Roel Goossens <roel@helpper.be>
Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>