refactor(tasks): always explicitly initialize Rayon thread pool (#13123)
Similar to #13122. Unconditionally initialize Rayon global thread pool in coverage runner.

The guarantees provided by explicitly initializing the thread pool are going to be critical for the linter.

Coverage currently doesn't cover linter, so it doesn't matter, but this doesn't hurt to do, and gives us greater safety if coverage does involve the linter at a later date.