refactor(linter): always explicitly initialize Rayon thread pool (#13122)
Previously we only explicitly initialized the global Rayon thread pool if `--threads` option is specified.

Instead, do it unconditionally, and always specify thread count. The purpose is to obtain a guarantee that `rayon::current_num_threads()` will always return the same number during the rest of the linter's work.

This invariant is not relied upon currently, but will become critical for soundness once we run JS plugins on multiple threads.

Also, if `std::thread::available_parallelism()` is not able to obtain CPU core count, log a warning that the linter will run single-threaded. Previously Rayon would use only a single thread, but would make that decision silently. I'm not sure in what circumstances `available_parallelism()` can return `Err`, but if it does, the user will probably want to know about it.