feat(allocator): add `data_end_ptr` and `set_cursor_ptr` methods to `Allocator` (#13134)
Add 2 more unsafe methods to `Allocator` to meddle with its internals. These methods are only present when `from_raw_parts` cargo feature is enabled.

These methods are used in a later PR in this stack, to write flags for which JS threads an `Allocator` is shared with into the `Allocator`s memory.