fix(allocator): fix soundness hole in `Allocator::alloc_bytes_start` (#13141)
Fix a bug in `Allocator::alloc_bytes_start` which could have resulted in UB.

Previously, if called with `size: usize::MAX`, the rounding up calculation would wrap around so it would not allocate any space in the arena. This PR makes sure such a large `size` causes a panic as it should.

In practice I don't think there was a real risk of `alloc_bytes_start` being called with such an absurdly large `size`, but regardless, we need to handle the edge cases.